+++ kodilla-cucumber/src/main/java/wallet/CashSlot.java
@@ -1,15 +1,16 @@
 package wallet;
 
 public class CashSlot {
-    private int contents;
+    private long contents;
 
-    public int getContents() {
+    public long getContents() {
         return contents;
     }
 
-    public void dispense(int amount) {
+    public void dispense(long amount) {
         if (amount <= 0) {
-            throw new IllegalArgumentException("Dispense amount must be positive, got: " + amount);
+            throw new IllegalArgumentException(
+                    "Dispense amount must be positive, got: " + amount);
         }
         this.contents = amount;
     }
